+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/ProjectSystem/RazorProjectService.cs
@@ -13,13 +13,13 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.ProjectSystem
     {
         public abstract void AddDocument(string filePath);
 
-        public abstract void OpenDocument(string filePath, SourceText sourceText, long version);
+        public abstract void OpenDocument(string filePath, SourceText sourceText, int version);
 
         public abstract void CloseDocument(string filePath);
 
         public abstract void RemoveDocument(string filePath);
 
-        public abstract void UpdateDocument(string filePath, SourceText sourceText, long version);
+        public abstract void UpdateDocument(string filePath, SourceText sourceText, int version);
 
         public abstract void AddProject(string filePath);
