+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/DefaultRazorProjectServiceTest.cs
@@ -669,8 +669,8 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer
             DocumentSnapshot documentSnapshot = TestDocumentSnapshot.Create(documentFilePath);
             var documentResolver = Mock.Of<DocumentResolver>(resolver => resolver.TryResolveDocument(It.IsAny<string>(), out documentSnapshot) == true);
             var documentVersionCache = new Mock<DocumentVersionCache>(MockBehavior.Strict);
-            documentVersionCache.Setup(cache => cache.TrackDocumentVersion(documentSnapshot, It.IsAny<long>()))
-                .Callback<DocumentSnapshot, long>((snapshot, version) =>
+            documentVersionCache.Setup(cache => cache.TrackDocumentVersion(documentSnapshot, It.IsAny<int>()))
+                .Callback<DocumentSnapshot, int>((snapshot, version) =>
                 {
                     Assert.Same(documentSnapshot, snapshot);
                     Assert.Equal(1337, version);
@@ -702,7 +702,7 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer
                 },
                 TestProjectSnapshot.Create("C:/__MISC_PROJECT__"));
             var documentVersionCache = new Mock<DocumentVersionCache>();
-            documentVersionCache.Setup(cache => cache.TrackDocumentVersion(It.IsAny<DocumentSnapshot>(), It.IsAny<long>()))
+            documentVersionCache.Setup(cache => cache.TrackDocumentVersion(It.IsAny<DocumentSnapshot>(), It.IsAny<int>()))
                 .Throws<XunitException>();
             var newText = SourceText.From("Something New");
             var projectService = CreateProjectService(
